Fix mismatching declaration/definition in disassembler.
Clone of https://codereview.chromium.org/2325683004/ (landing separately to fix the build).

TBR=loskutov@google.com

Review URL: https://codereview.chromium.org/2330453002 .

+++ syzygy/core/disassembler_util.cc
@@ -215,7 +215,7 @@ bool DecodeOneInstruction(uint32_t address,
   code.dt = Decode32Bits;
   code.features = DF_NONE;
   code.codeOffset = address;
-  code.codeLen = length;
+  code.codeLen = static_cast<int>(length);
   code.code = buffer;
 
   unsigned int decoded = 0;
